deffo needs polishing
added a javascript to the 'about me' page but it looks like its doesnt want to work-if you could figure out how please do so! I have also got rid of the back gorund color being grey in ALL the pages in the 'style css' as it was discrupting the other images.